wlan: Use Metainfo to fill privacy bit in the frame control
For WAI frames, gUcIsWai is used to decide the privacy bit in the
frame control in case tlState is WLANTL_STA_AUTHENTICATED.
gUcIsWai can be 0 for WAI frames in some case, and thus privacy
bit of WAI frame is set. So instead use Metainfo->ucIsWai, which
is filled for all WAI frames.

CRs-Fixed: 842163
Change-Id: I077b4de8783c7af2747a3d364ab1193da7cf8126