wlan: Fix excessive logging.
In vos_pkt_get_os_packet api, skb is made as NULL.
Latter on to check whether received packet is EAPOL,
hdd_IsEAPOLPacket api is called. In this api, if skb
is NULL, an error log is printed. Since skb is already
made as NULL, for every rx packet this log is getting
printed resulting in excessive logging. As a part of fix,
make sure that skb is freed only after packet is checked
for EAPOL.

Change-Id: I055ba1da5c2c4bc8d544b0b5b767a8634a36908d
CRs-Fixed: 849065